Use single-precision division for case 4 rowAnimFrame in DrawOptionMenu
The target divides m_optionRowAnim by FLOAT_8033362C in single
precision (fdivs) and converts to int inline via fctiwz, rather than
promoting to double and calling __cvt_fp2unsigned. Casting through int
instead of unsigned-via-double matches the target's conversion path.

main/MenuUtil 71.29 -> 71.43; DrawOptionMenu 54.09 -> 54.33.

+++ src/MenuUtil.cpp
@@ -1475,7 +1475,7 @@ void CMenuPcs::DrawOptionMenu()
 		if (static_cast<double>(m_optionRowAnim) >= DOUBLE_803335C0) {
 			rowAnimFrame = 0xD;
 		} else {
-			rowAnimFrame = static_cast<unsigned int>(static_cast<double>(m_optionRowAnim) / FLOAT_8033362C);
+			rowAnimFrame = static_cast<unsigned int>(static_cast<int>(m_optionRowAnim / FLOAT_8033362C));
 		}
 		const float specialRowCos = static_cast<float>(
 			cos(static_cast<double>(FLOAT_80333594 * static_cast<float>(rowAnimFrame) * FLOAT_80333630)));
